Always reset `Float64` as default float type at the end (#375)

+++ test/all_sky_radiative_transfer.jl
@@ -10,7 +10,7 @@ using ClimaComms
 using NCDatasets
 using RRTMGP
 
-@testset "All-sky full-spectrum RadiativeTransferModel [$FT]" for FT in (Float64, Float32)
+@testset "All-sky full-spectrum RadiativeTransferModel [$FT]" for FT in (Float32, Float64)
 
     @testset "Constructor argument errors" begin
         Oceananigans.defaults.FloatType = FT
@@ -120,7 +120,7 @@ using RRTMGP
         @test (lw_up_diff > 0) || (sw_dn_diff > 0)
     end
 
-    @testset "Custom effective radius models [$FT]" for FT in (Float64, Float32)
+    @testset "Custom effective radius models [$FT]" for FT in (Float32, Float64)
         Oceananigans.defaults.FloatType = FT
         topology = (Flat, Flat, Bounded)
         grid = RectilinearGrid(default_arch; size=8, x=0, y=45, z=(0, 10kilometers), topology)
